Initial checkin of Eto.Pcl.csproj. This builds Eto as a Portable Class Library. Eto builds but does not work yet. The reflection API's in particular, need to be moved to the PCL pattern. There are two build symbols at play:
PCL: Non-portable code that cannot not be made portable is disabled via #if !PCL
WINRT: Code that is not currently portable but which can be made portable is marked with this symbol. Once the code has been made portable, it should be marked with #if PCL, or just enabled everywhere (if Eto itself is upgraded to .net 4.5).