Wpf: Ensure proper sizing rules are followed for all controls
This simplifies the eto sizing rules for WPF by ensuring each container control overrides MeasureOveride() to provide custom logic. With this, we no longer need to manually calculate the scroll size for the Scrollable which was riddled with bugs.

This also fixes numerous issues where WPF wouldn’t respect the size set of a control in certain circumstances, which happened quite a bit after turning off the UseContentSize for all controls.