Mac: Fix Scrollable when dynamically added to a container with a background color
When using a background color (which required a layer), we used CanDrawSubviewsIntoLayer to collapse layers.  This unfortunately made the Scrollable draw its content on the containing layer initially when it was dynamically added to the form.

We no longer use layers to draw the background color - we add the drawRect: method to the ContainerControl dynamically and draw it there.