Wpf: fix occasional "'Hyperlink' is on a Visual a Visual3D" exception (#814)
Occasionally (I could not reproduce it in a simple snippet of code), I get a "'Hyperlink' is on a Visual a Visual3D" exception when doing a TreeGridView.ReloadData().
The fix consists in not looking for a parent if the control is of type Hyperlink when it is checked for the focus.