Wpf: Fix selection attributes when using a composition (IME)
When composing text using the Windows IME (e.g. for Japanese, Chinese, etc), it doesn't appear to use the selection attributes to enter the text.
This fixes it via a hack to update the attributes after the composition is finished.
Unfortunately, there doesn't seem to be a way to apply the attributes beforehand so when composing it uses the current attributes.