Wpf: Fix regressions with GridView click events with editable cells.
With the change in behaviour of the GridView to allow for drag/drop it broke the CellClick and CellDoubleClick events when the cell is editable.  Now the new behaviour takes into account how many clicks caused the cell to be edited and also will only start the editing if the cell already has focus to bring it back to previous/default behaviour.