Android platform building and running against latest Eto API. Many more controls implemented but still incomplete.
The only change outside of the Android projects for this commit is defining the Eto.Forms.ILayout interface, and copying definition of Eto.Forms.Layout.Update() to it. This allows StackLayout to implement ILayout so Update() becomes accessible, whereas previously there was no Update() because StackLayout does not actually derive from Layout. As this is only exposing a *new* and previously unused/unavailable method, this change will not affect any other platforms or existing code.