Unsubscribe from the input event handler
Class: 'Binding', method: 'RemovePropertyEvent'.
If we don't do that, dotMemory (and probably other
memory profiling tools) will report about an event
leakage of 'PropertyNotifyHelper'.
At the moment, nothing is done with the
EventHandler<EventArgs>, that is passed as an
input parameter. Performing the change fixes the
report of the memory profiling tool.