Fix an event handler leak in `Binding.RemovePropertyEvent`
Removing the handler in `RemovePropertyEvent` doesn't work, as the
`Target` of the handler is not the `PropertyNotifyHelper` instance; it
is the object that the handler is bound to.

Instead, find the correct handler to remove by checking all invocations
of the `PropertyChanged` event of the bound object, and comparing the
internal handler of the invocation with the method's argument.

Also, add unit tests to verify the correct behavior.